test: Fix smoke test for Hadoop 3.4.0
