--- tests/templates/kuttl/smoke/51-assert.yaml
@@ -1,5 +1,0 @@
----
-apiVersion: kuttl.dev/v1beta1
-kind: TestAssert
-commands:
-  - script: kubectl exec -n $NAMESPACE test-runner-0 -- python /tmp/metrics-test.py $NAMESPACE
