+++ tests/templates/kuttl/smoke/51-copy-metrics-test-script.yaml
@@ -2,4 +2,4 @@
 apiVersion: kuttl.dev/v1beta1
 kind: TestStep
 commands:
-  - script: kubectl cp -n $NAMESPACE ./metrics-test.py test-runner-0:/tmp
+  - script: kubectl cp -n $NAMESPACE ./test_metrics.py test-runner-0:/tmp
